# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing buildings and passes.",
        "title": "Service-propusk API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "172.17.17.145:8000",
    "basePath": "/",
    "paths": {
...
        "/passes/{id}/reject": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Меняет статус пропуска на \"rejected\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "passes"
                ],
                "summary": "Отклонить пропуск",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID пропуска (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/passhandler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/passhandler.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/passhandler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/passhandler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/passes/{id}/submit": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Меняет статус пропуска на \"submitted\" для рассмотрения модератором",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "passes"
                ],
                "summary": "Отправить пропуск на рассмотрение",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID пропуска (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/passhandler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/passhandler.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/passhandler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/passhandler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/passes/{passId}/building/{buildingId}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Удаляет здание из пропуска",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pass-buildings"
                ],
                "summary": "Удалить здание из пропуска",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID пропуска (UUID)",
                        "name": "passId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID здания (UUID)",
                        "name": "buildingId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/passBuildingHandler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/passBuildingHandler.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/passBuildingHandler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/passBuildingHandler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/passes/{passId}/building/{buildingId}/main": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Обновляет информацию о связи пропуска со зданием (комментарий)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pass-buildings"
                ],
                "summary": "Обновить связь пропуска со зданием",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID пропуска (UUID)",
                        "name": "passId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID здания (UUID)",
                        "name": "buildingId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Обновленные данные связи",
                        "name": "passBuilding",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/passBuildingHandler.EditPassBuildingRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/passBuildingHandler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/passBuildingHandler.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/passBuildingHandler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/passBuildingHandler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "passBuildingHandler.EditPassBuildingRequest": {
            "description": "Запрос на обновление информации о связи пропуска со зданием",
            "type": "object",
            "properties": {
                "comment": {
                    "description": "Комментарий к зданию в пропуске",
                    "type": "string",
                    "example": "Комментарий к зданию в пропуске"
                }
            }
        },
        "passBuildingHandler.ErrorResponse": {
            "description": "Структура ответа при возникновении ошибки",
            "type": "object",
            "properties": {
                "message": {
                    "description": "Текст ошибки",
                    "type": "string",
                    "example": "Описание ошибки"
                }
            }
        },
...
        "passhandler.ErrorResponse": {
            "description": "Структура ответа при возникновении ошибки",
            "type": "object",
            "properties": {
                "message": {
                    "description": "Текст ошибки",
                    "type": "string",
                    "example": "Описание ошибки"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "cookieAuth": {
            "description": "Enter your JWT token value",
            "type": "apiKey",
            "name": "X-Access-Token",
            "in": "header"
        }
    }
}